pass the data as props
