+++ src/components/config.js
@@ -1,0 +1,9 @@
+export const firebaseConfig = {
+  apiKey: "AIzaSyBl1k42_raOZuLIuexXbHoc-oV0nORdMb0",
+  authDomain: "info340-group4.firebaseapp.com",
+  databaseURL: "https://info340-group4-default-rtdb.firebaseio.com",
+  projectId: "info340-group4",
+  storageBucket: "info340-group4.appspot.com",
+  messagingSenderId: "396468489371",
+  appId: "1:396468489371:web:827b38bacf1e3abd60ca3e"
+};
